requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
datasets>=2.12.0
streamlit>=1.25.0
speechrecognition>=3.10.0
gtts>=2.3.0
pyaudio>=0.2.11
python-dotenv>=1.0.0
wandb>=0.15.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
